pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "behemoth-fuzzer"
version = "1.0.0"
description = "An AI-Agent API Security Fuzzer powered by Gemini 3"
dependencies = [
    "google-genai",
    "rich",
    "typer",
    "requests",
    "python-dotenv"
]

[project.scripts]
behemoth = "behemoth.cli:app"